v0.1.6: Region-specific STS tokens
With the recent launch of the ap-southeast-4 Melbourne region, I
noticed that tokens issued by `aws sts` will only work in default
regions unless a region-specific endpoint is used.

If you try to run a region-specific command in a non-default region
with a default token, you will get an error like this:

```
$ aws s3 ls

An error occurred (InvalidToken) when calling the ListBuckets operation: The provided token is malformed or otherwise invalid.

$ aws dynamodb create-table --region ap-southeast-4 ...

An error occurred (UnrecognizedClientException) when calling the CreateTable operation: The security token included in the request is invalid
```

With this changeset, calls to `aws sts` by `Update-AwsMfaCredentials`
will use the default region of the current profile, if configured, or
`us-east-1` if not configured.

It's worth noting that some people might want to have multiple profiles
tied to the same set of MFA credentials but using different default
regions. This isn't well supported by this changeset, so I might need
to investigate ways to handle this better in the future.